agentwatch-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
firebase-admin==6.5.0
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.2
pydantic==2.8.2
resend==2.3.0
openai==1.40.0
anthropic==0.34.0
groq>=0.4.0
